refactor socials
